requirements.txt
pandas
requests
sqlalchemy
openpyxl
pyxlsb
python-dotenv
psycopg2-binary